refactor: rename ToggleArchiveModal
